third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/stylenonce-allowed.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Security-Policy" content="style-src 'nonce-noncynonce'">
        <script src="/js-test-resources/js-test-pre.js"></script>
        <style nonce="noncynonce">
            #test {
                color: green;
            }
        </style>
        <style>
            #test {
                color: red;
            }
        </style>
    </head>
    <body>
        <p id="test">This text should be green.</p>
        <script>
            description("Style correctly whitelisted via a 'nonce-*' expression in 'style-src' should be applied to the page.");    
            var el = document.querySelector('#test');
            shouldBeEqualToString("window.getComputedStyle(el).color", "rgb(0, 128, 0)");
        </script>
        <script src="/js-test-resources/js-test-post.js"></script>
    </body>
</html>
