third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/stylenonce-allowed-expected.txt

CONSOLE MESSAGE: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'nonce-noncynonce'".

CONSOLE MESSAGE: line 11: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'nonce-noncynonce'".

Style correctly whitelisted via a 'nonce-*' expression in 'style-src' should be applied to the page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(el).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
This text should be green.
